refactor: consistent naming of files
